requirements.txt
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.2
importlib-metadata==5.2.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pycryptodome==3.16.0
pyrsistent==0.19.2
pytest==7.2.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
